nguoSystem/nguoSystem/backend/styles/serializers.py
from rest_framework import serializers
from .models import Style


class StyleSerializer(serializers.ModelSerializer):
    image_url = serializers.SerializerMethodField()

    class Meta:
        model = Style
        fields = [
            'id', 'name', 'description', 'price', 'gender',
            'category', 'image', 'image_url', 'created_at', 'updated_at'
        ]
        read_only_fields = ['id', 'created_at', 'updated_at']

    def get_image_url(self, obj):
        request = self.context.get('request')
        if obj.image and request:
            return request.build_absolute_uri(obj.image.url)
        return None
