nguoSystem/nguoSystem/backend/requirements.txt
# Core
Django==6.0.1
Pillow>=10.0.0
python-decouple==3.8
# REST API
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Production utilities
gunicorn==23.0.0
